fix: added changelog item
