+++ examples/default/ios/Podfile
@@ -26,7 +26,6 @@ target 'InstabugExample' do
   config = use_native_modules!
   rn_maps_path = '../node_modules/react-native-maps'
   pod 'react-native-google-maps', :path => rn_maps_path
-  pod 'Instabug', :podspec => 'https://ios-releases.instabug.com/custom/feature-w3c-header-cp-support/13.2.0/Instabug.podspec'
   # Flags change depending on the env values.
   flags = get_default_flags()
