+++ ios/native.rb
@@ -1,4 +1,4 @@
-$instabug = { :version => '13.2.0' }
+$instabug = { :version => '13.4.0' }
 
 def use_instabug! (spec = nil)
   version = $instabug[:version]
